pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "threat-research-mcp"
version = "0.2.0"
description = "Defensive security MCP server with multi-agent orchestration."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Harsh Thakur" }]
dependencies = [
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "mcp>=1.8.0; python_version >= '3.10'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.8.0",
  "mypy>=1.10.0",
  "bandit>=1.7.9",
  "pip-audit>=2.7.3",
  "build>=1.2.1",
  "twine>=5.1.1",
]

[project.scripts]
threat-research-mcp = "threat_research_mcp.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
